Fixed subcmd usage msg in `help`, made handle `myexe help sub cmd` form
1.  For main {Thor} `myexe` with subcommand `sub` containing command
`cmd`

        myexe sub help cmd

    was outputing something like

        Usage: sub cmd ...

    *without* the `myexe` part... which won't work if `sub` is only
callable through `myexe`.

    Maybe that's not how Thor's supposed to work, but that's how I'm
using it, so I fixed it.

2.  For that same setup, you can now run

        myexe help sub cmd

    and get the same thing as

        myexe sub help cmd

    because the first feels a lot more useful to me when exploring
commands because you just change the end.

3.  Trimmed the line lengths in the changed file to 80 columns 'cause
that's the width I use for tabs and it was really annoying to have
longer lines in there.

    In my opinion it makes sense to pick *something* for a max line
lenght so you don't need to adjust tab width all the time, and 80 has
always seemed like the natural choice.